requirements.txt
absl-py==2.0.0
albumentations==1.3.1
altair==5.1.2
appdirs==1.4.4
arrow==1.2.3
async-lru==2.0.4
brotlipy==0.7.0
cachetools==5.3.1
chardet==5.2.0
click==8.1.7
cloudpickle==3.0.0
cmake==3.27.2
contourpy==1.1.0
cycler==0.11.0
dask==2023.11.0
dask-jobqueue==0.8.2
distributed==2023.11.0
docker-pycreds==0.4.0
exceptiongroup==1.1.3
filelock==3.12.3
fonttools==4.42.1
fqdn==1.5.1
fsspec==2023.10.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.34
google-auth==2.23.3
google-auth-oauthlib==1.0.0
grpcio==1.59.0
imageio==2.31.1
importlib-metadata==6.8.0
isoduration==20.11.0
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema-specifications==2023.7.1
jupyter-lsp==2.2.0
jupyter_server_terminals==0.4.4
kaggle==1.5.16
kiwisolver==1.4.5
lazy_loader==0.3
lit==16.0.6
locket==1.0.0
Mako==1.2.4
Markdown==3.4.4
matplotlib==3.7.2
medcam==0.1.21
mpmath==1.3.0
msgpack==1.0.7
networkx==3.1
nibabel==5.1.0
numpy==1.23.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.8.0.76
opencv-python-headless==4.8.1.78
overrides==7.4.0
pandas==2.0.3
partd==1.4.1
pathtools==0.1.2
patsy==0.5.3
Pillow==10.0.0
plantcv==4.0.1
ply==3.11
protobuf==4.24.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
python-slugify==8.0.1
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
referencing==0.30.2
requests-oauthlib==1.3.1
rpds-py==0.10.2
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.13.0
sentry-sdk==1.30.0
setproctitle==1.3.2
SimpleITK==2.2.1
smmap==5.0.0
sortedcontainers==2.4.0
statsmodels==0.14.0
sympy==1.12
tblib==3.0.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
text-unidecode==1.3
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.8.12
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
tzdata==2023.3
ultralytics==8.0.207
uri-template==1.3.0
vl-convert-python==1.1.0
wandb==0.15.9
webcolors==1.13
webencodings==0.5.1
Werkzeug==3.0.0
xarray==2023.11.0
zict==3.0.0
zipp==3.17.0